starting the movement of db calls to separate files
